% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_edge_prediction_MSE.R
\name{conditional_edge_prediction_MSE}
\alias{conditional_edge_prediction_MSE}
\title{A Function to calculate Mean Edgewise MSE to evaluate edge predictions.}
\usage{
conditional_edge_prediction_MSE(edge_prediction_results)
}
\arguments{
\item{edge_prediction_results}{A list object returned by the
`conditional_edge_prediction()` function.}
}
\value{
A list of MSE's.
}
\description{
Calculates mean edgewise MSE for predicted edge values.
}
