% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_edge_prediction_plot.R
\name{conditional_edge_prediction_plot}
\alias{conditional_edge_prediction_plot}
\title{A Function generate edge prediction plots.}
\usage{
conditional_edge_prediction_plot(edge_prediction_results, filename, plot_size,
  node_name_cex = 1)
}
\arguments{
\item{edge_prediction_results}{A list object returned by the
`conditional_edge_prediction()` function.}

\item{filename}{The name of the file the user would like to save the plots in.
Should be something along the lines of "edge_plots.pdf"}

\item{plot_size}{A single number specifying the dimensions of the PDF output
file. This will automatically be square so only one number is required.}

\item{node_name_cex}{The cex for node names (printed on the diagonal).
Defaults to 1, but can be increased to make these easier to see.}
}
\value{
Saves a PDF plot to the current working directory
}
\description{
Plots edgewise predictions generated by the
`conditional_edge_prediction()` function.
}
