% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{covariate_data_2005}
\alias{covariate_data_2005}
\title{GDP and G8 membership data for 17 large countries from 2005.}
\format{A data frame with 17 rows and 3 columns}
\source{
\url{http://arxiv.org/abs/1505.04015}
}
\usage{
covariate_data_2005
}
\description{
Metadata for each country in "lending_2005", in cluding GDP, logged GDP, and
a G8 membership indicator.
}
\keyword{datasets}

