% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_example_simulated_networks.R
\name{find_example_simulated_network}
\alias{find_example_simulated_network}
\title{Find an example simulated network that is the minimum Frobenius  distance from the observed network.}
\usage{
find_example_simulated_network(GERGM_Object, observed_network,
  collapse_to_correlation_space = F)
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the
GERGM function.}

\item{observed_network}{The observed network used as the dependent variable
in the original gergm() specification.}

\item{collapse_to_correlation_space}{If TRUE, then all simulated networks are
normalized so the maximum absolute edge value is less than one. Defaults to
FALSE.}
}
\value{
A simulated network (as a matrix object).
}
\description{
Find an example simulated network that is the minimum Frobenius  distance from the observed network.
}
