% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{GOF}
\alias{GOF}
\title{Generate Goodness Of Fit plot from a GERGM object.}
\usage{
GOF(GERGM_Object, ...)
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the
GERGM function.}

\item{...}{Additional Arguments can be passed in. Included for eventual compatibility with XERGM package.}
}
\value{
A set of box plots where of simulated network statistics centered at the observed value for those statistics and normalized by their standard deviation. This aids in interpretation as the y-axis can be interpreted as the number of simulated-sample standard deviations above or below the observed statistic.
}
\description{
Generate Goodness Of Fit plot from a GERGM object.
}

