% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{lending_2005}
\alias{lending_2005}
\title{Logged international lending volumes for 17 large countries from 2005.}
\format{A matrix with 17 rows and columns and a zero diagonal.}
\source{
\url{http://arxiv.org/abs/1505.04015}
}
\usage{
lending_2005
}
\description{
A dataset of logged aggregate public and private lending volumes between 17
large industrialized countries.
}
\keyword{datasets}

