% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{net_exports_2005}
\alias{net_exports_2005}
\title{Normalized net exports between 17 large countries from 2005.}
\format{A matrix with 17 rows and columns and a zero diagonal.}
\source{
\url{http://arxiv.org/abs/1505.04015}
}
\usage{
net_exports_2005
}
\description{
Aggregate net export volume between countries has been normalized to lie
between 0 and 1.
}
\keyword{datasets}
