% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trace_Plot.R
\name{Trace_Plot}
\alias{Trace_Plot}
\title{Generate trace plot of network density from a GERGM object.}
\usage{
Trace_Plot(GERGM_Object)
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the
GERGM function.}
}
\value{
A trace plot of network density.
}
\description{
Generate trace plot of network density from a GERGM object.
}
