% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_edge_prediction_correlation_plot.R
\name{conditional_edge_prediction_correlation_plot}
\alias{conditional_edge_prediction_correlation_plot}
\title{A Function generate a scatter plot of observed vs predicted edge values.}
\usage{
conditional_edge_prediction_correlation_plot(edge_prediction_results,
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{edge_prediction_results}{A list object returned by the
`conditional_edge_prediction()` function.}

\item{xlim}{Defaults to NULL, but can be set by the user if desired to a
numeric vector of the form c(lower,upper) for observed edge values.}

\item{ylim}{Defaults to NULL, but can be set by the user if desired to a
numeric vector of the form c(lower,upper) for predicted edge values.}
}
\value{
Saves a PDF plot to the current working directory
}
\description{
Scatter plot of observed vs. predicted edge values generated by the
`conditional_edge_prediction()` function. Also calculated correlation coefficient.
}

