% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Plot.R
\name{Estimate_Plot}
\alias{Estimate_Plot}
\title{Generate parameter estimate plot with 95 percent CI's from a GERGM object.}
\usage{
Estimate_Plot(GERGM_Object, normalize_coefficients = FALSE,
  coefficients_to_plot = c("both", "covariate", "structural"))
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the
GERGM function.}

\item{normalize_coefficients}{Defaults to FALSE, if TRUE then parameter
estimates will be converted be deivided by their standard deviations with
and displayed with 95 percent confidence intervals. These coefficients will
no longer be comparable, but make graphical interpretation of significance
and sign easier.}

\item{coefficients_to_plot}{An optional argument indicating which kind of
parameters to plot. Can be one of "both","covariate", or "structural". Useful
for creating separate parameter plots for covariates and structural
parameters when these parameters are on very different scales.}
}
\value{
A parameter estimate plot.
}
\description{
Generate parameter estimate plot with 95 percent CI's from a GERGM object.
}

