% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Estimate_Plot.R
\name{Estimate_Plot}
\alias{Estimate_Plot}
\title{Generate parameter estimate plot with 95 percent CI's from a GERGM object.}
\usage{
Estimate_Plot(GERGM_Object)
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the GERGM function.}
}
\value{
A parameter estimate plot.
}
\description{
Generate parameter estimate plot with 95 percent CI's from a GERGM object.
}

