% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{GOF}
\alias{GOF}
\title{Generate Goodness Of Fit plot from a GERGM object.}
\usage{
GOF(GERGM_Object)
}
\arguments{
\item{GERGM_Object}{The object returned by the estimation procedure using the GERGM function.}
}
\value{
A GOF plot.
}
\description{
Generate Goodness Of Fit plot from a GERGM object.
}

