% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysteresis_plot.R
\name{hysteresis_plot}
\alias{hysteresis_plot}
\title{Generate hysteresis plots for theta parameter estimates}
\usage{
hysteresis_plot(hysteresis_output, text_size = 12, ...)
}
\arguments{
\item{hysteresis_output}{The list object output from the hysteresis function.}

\item{text_size}{The base size for axis text. Defaults to 12.}

\item{...}{Additional arguments currently not supported.}
}
\description{
Generate hysteresis plots for theta parameter estimates
}
